extract reusable parts of BoxComponentMode
